requirement.txt
absl-py==2.3.1
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==14.0.1
bitsandbytes==0.46.1
blake3==1.0.5
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.8.1
contourpy==1.3.2
cupy-cuda12x==13.5.1
cycler==0.12.1
datasets==3.1.0
deepspeed==0.16.2
depyf==0.18.0
descartes==1.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastrlock==0.8.3
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.9.0
future==1.0.0
gguf==0.10.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
interegular==0.3.3
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsondiff==2.2.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
liger_kernel==0.5.0
llamafactory==0.9.2.dev0
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lmdb==1.6.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
mistral_common==1.8.1
modelscope==1.22.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.6
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post6
peft==0.12.0
pillow==10.4.0
pip==25.1
pluggy==1.6.0
prometheus_client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.2
protobuf==6.31.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybind11==3.0.0
pycocotools==2.0.10
pycountry==24.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-extra-types==2.10.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pyquaternion==0.9.9
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==1.2.10
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rouge-chinese==1.0.3
rpds-py==0.26.0
ruff==0.12.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.33.0
setuptools==76.0.0
Shapely==1.8.5.post1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==0.2.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.2
triton==3.1.0
trl==0.9.6
typer==0.16.0
typing_extensions==4.12.2
typing-inspection==0.4.1
tyro==0.8.14
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.6.5
watchfiles==1.1.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
